pinned-items-double-click: Fix a small, harmless typo

+++ mods/pinned-items-double-click.wh.cpp
@@ -114,7 +114,7 @@ void WINAPI CTaskListWnd__HandleClick_Hook(PVOID pThis,
     static PVOID firstClickTaskBtnGroup;
 
     ULONGLONG tickCount = GetTickCount64();
-    ULONGLONG messagePos = GetMessagePos();
+    DWORD messagePos = GetMessagePos();
 
     if (firstClickTickCount &&
         IsDoubleClickDistance(firstClickMessagePos, messagePos) &&
